third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/navigate-same-document_currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for same-document navigation.navigate()
  assert_array_equals: expected property 2 to be "navigatesuccess" but got "committed fulfilled" (expected array ["navigate", "currententrychange", "navigatesuccess", "committed fulfilled", "promise microtask", "finished fulfilled"] got ["navigate", "currententrychange", "committed fulfilled", "promise microtask", "navigatesuccess", "finished fulfilled"])
Harness: the test ran to completion.
